requirements.txt
streamlit
python-dotenv
langchain
streamlit-extras
faiss-cpu
langchain-community
pandas
PyPDF2
transformers
scikit-learn
torch
requests
sentence_transformers
langchain_openai

langchain-google-genai

